+++ web/submission/views.py
@@ -188,8 +188,8 @@ def index(request, task_id=None, resubmit_hash=None):
         if request.POST.get("import_reconstruction"):
             options += "import_reconstruction=1,"
 
-        if request.POST.get("disable_cape"):
-            options += "disable_cape=1,"
+        if request.POST.get("unpacker"):
+            options += "unpacker=2,"
 
         if request.POST.get("kernel_analysis"):
             options += "kernel_analysis=yes,"